DEV: Update plugin to match latest guidelines
This commit updates the plugin to follow the latest guidelines, as
shown in discourse-plugin-skeleton, removing the explicit require calls,
using proper namespaces and using config/routes.rb for defining all
routes.
