+++ lib/discourse_gamification/engine.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module ::DiscourseGamification
+  class Engine < ::Rails::Engine
+    engine_name PLUGIN_NAME
+    isolate_namespace DiscourseGamification
+  end
+end
