+++ lib/discourse_gamification/guardian_extension.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-module DiscourseGamification
-  module Guardian
+module ::DiscourseGamification
+  module GuardianExtension
     def can_see_leaderboard?(leaderboard)
       return true if leaderboard.visible_to_groups_ids.empty?
       return true if self.is_admin?
