+++ lib/discourse_gamification/scorables/scorable.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-module DiscourseGamification
+module ::DiscourseGamification
   class Scorable
     class << self
       def enabled?
