+++ lib/discourse_gamification/scorables/user_invited.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module DiscourseGamification
+module ::DiscourseGamification
   class UserInvited < Scorable
     def self.score_multiplier
       SiteSetting.user_invited_score_value
