+++ lib/discourse_gamification/user_extension.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
-module DiscourseGamification
+
+module ::DiscourseGamification
   module UserExtension
     def gamification_score
       DiscourseGamification::GamificationScore.where(user_id: self.id).sum(:score)
